# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda API",
        "description": "Documentação da API de agendamentos e lembretes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://agenda-api.test",
            "description": "Servidor local de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/v1/reminders": {
            "get": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Listar todos os lembretes do usuário autenticado",
                "operationId": "ac26ac2ad062912d8d4f74dd2cd64704",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de lembretes por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lembretes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReminderResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Criar um novo lembrete",
                "operationId": "1b0c7dc52c3c59b0d7d16abb5baf2688",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReminderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lembrete criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReminderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reminders/{id}": {
            "get": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Visualizar um lembrete específico",
                "operationId": "95740e8f1b35d8d18a93555f0466aada",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do lembrete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReminderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Atualizar um lembrete existente",
                "operationId": "bec0a9feb2b9a7c09c74177876e1a7f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReminderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lembrete atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReminderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Remover um lembrete",
                "operationId": "cd450624baff8123ac5085947e880d49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lembrete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lembrete excluído com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReminderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "appointment_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "title": {
                        "type": "string",
                        "example": "Tomar medicamento"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomar remédio para dor"
                    },
                    "reminder_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-01T16:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "StoreReminderRequest": {
                "required": [
                    "title",
                    "description",
                    "reminder_time"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Tomar medicamento"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomar remédio para dor às 16h"
                    },
                    "reminder_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-01T16:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token de autenticação Bearer aqui",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}